backend/requirements.txt
fastapi
uvicorn
numpy
pydantichttpx==0.28.1
